fix warnings on mac

+++ Source/ConnectionGraph2.cpp
@@ -276,7 +276,7 @@ PluginReceiveResult ConnectionGraph2::OnReceive(Packet *packet)
 			bs.Read(saag.guid);
 			unsigned long idx2 = remoteSystems[idx]->remoteConnections.GetIndexFromKey(saag, &objectExists);
 			if (objectExists)
-				remoteSystems[idx]->remoteConnections.RemoveAtIndex(idx2);
+				remoteSystems[idx]->remoteConnections.RemoveAtIndex((unsigned int)idx2);
 		}
 	}
 	else if (packet->data[0]==ID_REMOTE_NEW_INCOMING_CONNECTION)
